fix(gridpanel): address final-review findings
- docs/controls/gridpanel.md: rewrite for the new model — designer-
  droppable TTyGridCell containers + ColumnSizes/RowSizes strings +
  weighted star + per-cell Padding; drop the deleted SetCell/
  SetColumnStyle API and the old "star Value ignored" semantics.
- antdesign dashboard: anchor GridKPI/GridMain [akLeft,akTop,akRight]
  so they actually widen with the window (the migration had dropped
  the old akRight responsiveness).
- TyGridTrackSizes: Int64 intermediate in the weighted-star multiply
  to harden against a pathological ColumnSizes weight overflowing.
- containers: drop a leftover blank line where WireGrid was removed.

+++ source/tyControls.GridPanel.pas
@@ -204,7 +204,9 @@ begin
         begin
           w := ATracks[i].Value;
           if w < 1 then w := 1;
-          Result[i] := (pool * w) div starWeight;
+          // Int64 intermediate: a pathological ColumnSizes weight (e.g. '200000*') on a
+          // wide axis would overflow a 32-bit multiply. The result itself fits in Integer.
+          Result[i] := Integer((Int64(pool) * w) div starWeight);
           Inc(given, Result[i]);
         end;
         if Result[i] < 0 then Result[i] := 0;
